+++ R/columns.R
@@ -1,4 +1,4 @@
-#' Create columns definition from data
+#' Create column definitions from data
 #' @param data (data.frame) data
 #' @param editor (bool): Whether to make columns editable.
 #' @param filter (bool): Whether to add a header filter to the columns.
@@ -107,7 +107,7 @@ set_formatter_money <- function(
     column,
     decimal = c(",", "."),
     thousand = c(".", ","),
-    symbol = "EUR",
+    symbol = "$", # "\U20AC"
     symbol_after = "p",
     negative_sign = "-",
     precision = FALSE,
@@ -211,7 +211,6 @@ set_formatter_star <- function(widget, column, number_of_stars, hoz_align = "cen
 #'  If set to \code{NA} the minimum value of the column is used.
 #' @param max (numeric): The maximum value for progress bar.
 #'  If set to \code{NA} the maximum value of the column is used.
-#' @param max description
 #' @param color (character): Either a single color or a vector of colors
 #' @param legend (character, \code{TRUE}, JavaScript function): If set to \code{TRUE} the value of the cell is displayed.
 #'  Set to \code{NA} to display no value at all.
@@ -300,7 +299,7 @@ set_formatter_toggle_switch <- function(
   modify_col_def(widget, column, col_update)
 }
 
-#' Datetime formatter
+#' Datetime Formatter
 #' @inheritParams set_formatter_html
 #' @param input_format (character): The datetime input format.
 #' @param output_format (character): The datetime output format.
@@ -314,8 +313,7 @@ set_formatter_datetime <- function(
     input_format = "yyyy-MM-dd hh:ss:mm",
     output_format = "yy/MM/dd",
     invalid_placeholder = "(invalid datetime)",
-    timezone = NA
-    ) {
+    timezone = NA) {
   # Body
   col_update <- list(
     formatter = "datetime",
@@ -340,6 +338,7 @@ set_formatter_color <- function(widget, column) {
 
 #' Traffic Light Formatter
 #' @inheritParams set_formatter_progress
+#' @example examples/formatters/formatter_traffic_light.R
 #' @export
 set_formatter_traffic_light <- function(
     widget,
@@ -349,9 +348,9 @@ set_formatter_traffic_light <- function(
     color = c("green", "orange", "red"),
     hoz_align = "center") {
   # Body
-  if (is.na(min)) min = min(widget$x$data[column])
+  if (is.na(min)) min <- min(widget$x$data[column])
 
-  if (is.na(max)) max = max(widget$x$data[column])
+  if (is.na(max)) max <- max(widget$x$data[column])
 
   col_update <- list(
     formatter = "traffic",
@@ -364,6 +363,44 @@ set_formatter_traffic_light <- function(
   )
   modify_col_def(widget, column, col_update)
 }
+
+# Other
+
+#' Make columns editable
+#' @inheritParams set_formatter_html
+#' @param columns (character vector): Columns the editor is applied to.
+#' @param type (character): Either \code{input} or \code{number}.
+#' @example examples/formatters/column_editor.R
+#' @export
+set_column_editor <- function(widget, columns, type = c("input", "number")) {
+  col_update <- list(editor = match.arg(type))
+  for (column in columns) {
+    widget <- modify_col_def(widget, column, col_update)
+  }
+
+  return(widget)
+}
+
+#' Add header filter
+#' @inheritParams set_column_editor
+#' @param columns (character vector): Columns the editor is applied to.
+#'  If set to \code{NULL}, the editor is applied to all columns.
+#' @example examples/misc/header_filter.R
+#' @export
+set_header_filter <- function(widget, columns = NULL) {
+  if (is.null(columns)) {
+    columns <- colnames(widget$x$data)
+  }
+
+  col_update <- list(headerFilter = TRUE)
+  for (column in columns) {
+    widget <- modify_col_def(widget, column, col_update)
+  }
+
+  return(widget)
+}
+
+# Generics
 
 modify_col_def <- function(widget, column, col_update) {
   for (index in 1:length(widget$x$options$columns)) {