+++ examples/formatters/formatter_html.R
@@ -11,5 +11,5 @@ data <- data.frame(
   )
 )
 
-tabulator(data, width=200) |>
+tabulator(data, width = 400) |>
   set_formatter_html("text_style")