+++ examples/formatters/formatter_textarea.R
@@ -3,7 +3,7 @@ data <- data.frame(
   txt = c(
     "This\nis\nsome\ntext\nwith\nmultiple\nline\nbreaks",
     "- R\n- Python\n- Julia"
-    )
+  )
 )
 
 tabulator(data, width = 200) |>