+++ examples/formatters/formatter_tick_cross.R
@@ -2,7 +2,7 @@ data <- data.frame(
   Artist = c("Art Blackey", "Nirvana", "Bob Marley"),
   Grunge = c(0, 1, 0),
   Jazz = c(1, 0, 0),
-  Reggae = c(0,0,1)
+  Reggae = c(0, 0, 1)
 )
 
 tabulator(data, width = 400) |>