+++ NEWS.md
@@ -1,3 +1,12 @@
+# rtabulator 0.1.1
+
+* Update docs
+* Add pkgdown site
+* Add function to create multi column headers
+* Add a lot of examples
+* Update README
+* Make `devtools::check()` pass 🚀
+
 # rtabulator 0.1.0
 
 * Add NEWS.md to track changes.